refactor components into feature-based subfolders
