+++ src/components/diagnostics/StartupDiagnostics.tsx
