+++ src/components/registration/PasswordStrengthMeter.tsx
