+++ src/main.tsx
@@ -1,11 +1,11 @@
 import { createRoot } from "react-dom/client";
 import { browserEnv } from "./lib/env";
-import StartupDiagnostics from "./components/StartupDiagnostics";
+import StartupDiagnostics from "./components/diagnostics/StartupDiagnostics.tsx";
 import "./index.css";
 
 import ErrorBoundary from "./components/common/ErrorBoundary.tsx";
 
-import { ThemeProvider } from "./components/theme-provider";
+import { ThemeProvider } from "./components/theme/theme-provider.tsx";
 
 
 const root = createRoot(document.getElementById("root")!);
